fmdb/requirements.txt
yfinance>=0.2.40
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
requests>=2.31
pandas-ta>=0.3.14b
